fixed errors and reformatted the District.cs Class file.
